refactor: move parseFileMetadata to separate file so it can be unit tested
